实现hessian协议上传大文件

+++ demo-up_download-provider/src/main/java/com/wuwei/dubboProvider/service/impl/UploadDemoServiceImpl.java
@@ -5,22 +5,29 @@ import com.wuwei.dubboApi.service.UploadDemoService;
 import org.apache.dubbo.config.annotation.Service;
 
 import java.io.*;
-import java.util.UUID;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 
-@Service
+@Service(protocol = "hessian")
 public class UploadDemoServiceImpl implements UploadDemoService {
 
 
     @Override
     public void uploadDocument(Document document, InputStream in) {
         String name = document.getName();
-        String uuid = UUID.randomUUID().toString();
+        String uuid = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyyMMddHHmmss"));
 
-        BufferedInputStream bufferedInputStream = new BufferedInputStream(in);
+        File uploadFileDirectory = new File(System.getProperty("user.dir") + File.separator + "upload/demo");
+        if (!uploadFileDirectory.exists() && !uploadFileDirectory.mkdirs()) {
+            System.out.println("创建失败");
+            return;
+        }
+
         String fileName = uuid + name;
-        File uploadFile = new File(System.getProperty("user.dir") + File.separator + "upload/demo" + File.separator + fileName);
+        File uploadFile = new File(uploadFileDirectory, fileName);
         try (BufferedOutputStream bufferedOutputStream = new BufferedOutputStream(new FileOutputStream(uploadFile));) {
             int i;
+            BufferedInputStream bufferedInputStream = new BufferedInputStream(in);
             while ((i = bufferedInputStream.read()) != -1) {
                 bufferedOutputStream.write(i);
             }
@@ -28,6 +35,8 @@ public class UploadDemoServiceImpl implements UploadDemoService {
         } catch (IOException e) {
             e.printStackTrace();
             System.out.println("***************上传失败！****************");
+        } catch (Exception ex) {
+            ex.printStackTrace();
         }
     }
 }
